Restructure hibernate integration
