--- dd-java-agent/instrumentation/hibernate-3.5/src/main/java/datadog/trace/instrumentation/hibernate/v3_5/CriteriaInstrumentation.java
@@ -1,83 +1,0 @@
-package datadog.trace.instrumentation.hibernate.v3_5;
-
-import static datadog.trace.agent.tooling.ByteBuddyElementMatchers.safeHasSuperType;
-import static java.util.Collections.singletonMap;
-import static net.bytebuddy.matcher.ElementMatchers.isInterface;
-import static net.bytebuddy.matcher.ElementMatchers.isMethod;
-import static net.bytebuddy.matcher.ElementMatchers.named;
-import static net.bytebuddy.matcher.ElementMatchers.not;
-
-import com.google.auto.service.AutoService;
-import datadog.trace.agent.tooling.Instrumenter;
-import datadog.trace.bootstrap.ContextStore;
-import datadog.trace.bootstrap.InstrumentationContext;
-import datadog.trace.instrumentation.hibernate.common.SessionMethodUtils;
-import datadog.trace.instrumentation.hibernate.common.SessionState;
-import java.util.Map;
-import net.bytebuddy.asm.Advice;
-import net.bytebuddy.description.method.MethodDescription;
-import net.bytebuddy.description.type.TypeDescription;
-import net.bytebuddy.implementation.bytecode.assign.Assigner;
-import net.bytebuddy.matcher.ElementMatcher;
-import org.hibernate.Criteria;
-
-@AutoService(Instrumenter.class)
-public class CriteriaInstrumentation extends Instrumenter.Default {
-
-  public CriteriaInstrumentation() {
-    super("hibernate", "hibernate-core");
-  }
-
-  @Override
-  public Map<String, String> contextStore() {
-    return singletonMap("org.hibernate.Criteria", SessionState.class.getName());
-  }
-
-  @Override
-  public String[] helperClassNames() {
-    return new String[] {
-      "datadog.trace.instrumentation.hibernate.common.SessionMethodUtils",
-      "datadog.trace.instrumentation.hibernate.common.SessionState",
-      "datadog.trace.agent.decorator.BaseDecorator",
-      "datadog.trace.agent.decorator.ClientDecorator",
-      "datadog.trace.agent.decorator.DatabaseClientDecorator",
-      "datadog.trace.agent.decorator.OrmClientDecorator",
-      "datadog.trace.instrumentation.hibernate.common.HibernateDecorator",
-    };
-  }
-
-  @Override
-  public ElementMatcher<TypeDescription> typeMatcher() {
-    return not(isInterface()).and(safeHasSuperType(named("org.hibernate.Criteria")));
-  }
-
-  @Override
-  public Map<? extends ElementMatcher<? super MethodDescription>, String> transformers() {
-    return singletonMap(
-        isMethod().and(named("list").or(named("uniqueResult")).or(named("scroll"))),
-        CriteriaMethodAdvice.class.getName());
-  }
-
-  public static class CriteriaMethodAdvice {
-
-    @Advice.OnMethodEnter(suppress = Throwable.class)
-    public static SessionState startMethod(
-        @Advice.This final Criteria criteria, @Advice.Origin("#m") final String name) {
-
-      final ContextStore<Criteria, SessionState> contextStore =
-          InstrumentationContext.get(Criteria.class, SessionState.class);
-
-      return SessionMethodUtils.startScopeFrom(
-          contextStore, criteria, "hibernate.criteria." + name, null, true);
-    }
-
-    @Advice.OnMethodExit(onThrowable = Throwable.class, suppress = Throwable.class)
-    public static void endMethod(
-        @Advice.Enter final SessionState state,
-        @Advice.Thrown final Throwable throwable,
-        @Advice.Return(typing = Assigner.Typing.DYNAMIC) final Object entity) {
-
-      SessionMethodUtils.closeScope(state, throwable, entity);
-    }
-  }
-}
