--- dd-java-agent/instrumentation/hibernate-4.3/hibernate-4.3.gradle
@@ -1,45 +1,0 @@
-// Set properties before any plugins get loaded
-project.ext {
-  // Execute tests on all JVMs, even rare and outdated ones
-  coreJavaInstrumentation = true
-  minJavaVersionForTests = JavaVersion.VERSION_1_7
-}
-
-muzzle {
-  pass {
-    group = "org.hibernate"
-    module = "hibernate-core"
-    versions = "[4.3.0.Final,)"
-  }
-}
-
-apply from: "${rootDir}/gradle/java.gradle"
-
-apply plugin: 'org.unbroken-dome.test-sets'
-
-testSets {
-  latestDepTest {
-    dirName = 'test'
-  }
-}
-
-dependencies {
-  compileOnly group: 'org.hibernate', name: 'hibernate-core', version: '4.3.0.Final'
-
-  compile project(':dd-java-agent:agent-tooling')
-  compile project(':dd-java-agent:instrumentation:hibernate-4.0')
-
-  compile deps.bytebuddy
-  compile deps.opentracing
-  annotationProcessor deps.autoservice
-  implementation deps.autoservice
-
-  testCompile project(':dd-java-agent:testing')
-  testCompile project(':dd-java-agent:instrumentation:jdbc')
-
-  testCompile group: 'org.hibernate', name: 'hibernate-core', version: '4.3.0.Final'
-  testCompile group: 'org.hsqldb', name: 'hsqldb', version: '2.0.0'
-
-  latestDepTestCompile group: 'org.hibernate', name: 'hibernate-core', version: '+'
-  latestDepTestCompile group: 'org.hsqldb', name: 'hsqldb', version: '2.0.0'
-}
