--- dd-java-agent/instrumentation/hibernate-4.3/src/main/java/datadog/trace/instrumentation/hibernate/v4_3/ProcedureCallInstrumentation.java
@@ -1,80 +1,0 @@
-package datadog.trace.instrumentation.hibernate.v4_3;
-
-import static datadog.trace.agent.tooling.ByteBuddyElementMatchers.safeHasSuperType;
-import static java.util.Collections.singletonMap;
-import static net.bytebuddy.matcher.ElementMatchers.isInterface;
-import static net.bytebuddy.matcher.ElementMatchers.isMethod;
-import static net.bytebuddy.matcher.ElementMatchers.named;
-import static net.bytebuddy.matcher.ElementMatchers.not;
-
-import com.google.auto.service.AutoService;
-import datadog.trace.agent.tooling.Instrumenter;
-import datadog.trace.bootstrap.ContextStore;
-import datadog.trace.bootstrap.InstrumentationContext;
-import datadog.trace.instrumentation.hibernate.common.SessionMethodUtils;
-import datadog.trace.instrumentation.hibernate.common.SessionState;
-import java.util.Map;
-import net.bytebuddy.asm.Advice;
-import net.bytebuddy.description.method.MethodDescription;
-import net.bytebuddy.description.type.TypeDescription;
-import net.bytebuddy.matcher.ElementMatcher;
-import org.hibernate.procedure.ProcedureCall;
-
-@AutoService(Instrumenter.class)
-public class ProcedureCallInstrumentation extends Instrumenter.Default {
-
-  public ProcedureCallInstrumentation() {
-    super("hibernate", "hibernate-core");
-  }
-
-  @Override
-  public Map<String, String> contextStore() {
-    return singletonMap("org.hibernate.procedure.ProcedureCall", SessionState.class.getName());
-  }
-
-  @Override
-  public String[] helperClassNames() {
-    return new String[] {
-      "datadog.trace.instrumentation.hibernate.common.SessionMethodUtils",
-      "datadog.trace.instrumentation.hibernate.common.SessionState",
-      "datadog.trace.agent.decorator.BaseDecorator",
-      "datadog.trace.agent.decorator.ClientDecorator",
-      "datadog.trace.agent.decorator.DatabaseClientDecorator",
-      "datadog.trace.agent.decorator.OrmClientDecorator",
-      "datadog.trace.instrumentation.hibernate.common.HibernateDecorator",
-    };
-  }
-
-  @Override
-  public ElementMatcher<TypeDescription> typeMatcher() {
-    return not(isInterface()).and(safeHasSuperType(named("org.hibernate.procedure.ProcedureCall")));
-  }
-
-  @Override
-  public Map<? extends ElementMatcher<? super MethodDescription>, String> transformers() {
-    return singletonMap(
-        isMethod().and(named("getOutputs")), ProcedureCallMethodAdvice.class.getName());
-  }
-
-  public static class ProcedureCallMethodAdvice {
-
-    @Advice.OnMethodEnter(suppress = Throwable.class)
-    public static SessionState startMethod(
-        @Advice.This final ProcedureCall call, @Advice.Origin("#m") final String name) {
-
-      final ContextStore<ProcedureCall, SessionState> contextStore =
-          InstrumentationContext.get(ProcedureCall.class, SessionState.class);
-
-      final SessionState state =
-          SessionMethodUtils.startScopeFrom(
-              contextStore, call, "hibernate.procedure." + name, call.getProcedureName(), true);
-      return state;
-    }
-
-    @Advice.OnMethodExit(onThrowable = Throwable.class, suppress = Throwable.class)
-    public static void endMethod(
-        @Advice.Enter final SessionState state, @Advice.Thrown final Throwable throwable) {
-      SessionMethodUtils.closeScope(state, throwable, null);
-    }
-  }
-}
