--- dd-java-agent/instrumentation/hibernate-4.3/src/main/java/datadog/trace/instrumentation/hibernate/v4_3/SessionInstrumentation.java
@@ -1,86 +1,0 @@
-package datadog.trace.instrumentation.hibernate.v4_3;
-
-import static datadog.trace.agent.tooling.ByteBuddyElementMatchers.safeHasSuperType;
-import static net.bytebuddy.matcher.ElementMatchers.isInterface;
-import static net.bytebuddy.matcher.ElementMatchers.isMethod;
-import static net.bytebuddy.matcher.ElementMatchers.named;
-import static net.bytebuddy.matcher.ElementMatchers.not;
-import static net.bytebuddy.matcher.ElementMatchers.returns;
-
-import com.google.auto.service.AutoService;
-import datadog.trace.agent.tooling.Instrumenter;
-import datadog.trace.bootstrap.ContextStore;
-import datadog.trace.bootstrap.InstrumentationContext;
-import datadog.trace.instrumentation.hibernate.common.SessionMethodUtils;
-import datadog.trace.instrumentation.hibernate.common.SessionState;
-import java.util.Collections;
-import java.util.HashMap;
-import java.util.Map;
-import net.bytebuddy.asm.Advice;
-import net.bytebuddy.description.method.MethodDescription;
-import net.bytebuddy.description.type.TypeDescription;
-import net.bytebuddy.matcher.ElementMatcher;
-import org.hibernate.SharedSessionContract;
-import org.hibernate.procedure.ProcedureCall;
-
-@AutoService(Instrumenter.class)
-public class SessionInstrumentation extends Instrumenter.Default {
-
-  public SessionInstrumentation() {
-    super("hibernate", "hibernate-core");
-  }
-
-  @Override
-  public Map<String, String> contextStore() {
-    final Map<String, String> map = new HashMap<>();
-    map.put("org.hibernate.SharedSessionContract", SessionState.class.getName());
-    map.put("org.hibernate.procedure.ProcedureCall", SessionState.class.getName());
-    return Collections.unmodifiableMap(map);
-  }
-
-  @Override
-  public String[] helperClassNames() {
-    return new String[] {
-      "datadog.trace.instrumentation.hibernate.common.SessionMethodUtils",
-      "datadog.trace.instrumentation.hibernate.common.SessionState",
-      "datadog.trace.agent.decorator.BaseDecorator",
-      "datadog.trace.agent.decorator.ClientDecorator",
-      "datadog.trace.agent.decorator.DatabaseClientDecorator",
-      "datadog.trace.agent.decorator.OrmClientDecorator",
-      "datadog.trace.instrumentation.hibernate.common.HibernateDecorator",
-    };
-  }
-
-  @Override
-  public ElementMatcher<TypeDescription> typeMatcher() {
-    return not(isInterface()).and(safeHasSuperType(named("org.hibernate.SharedSessionContract")));
-  }
-
-  @Override
-  public Map<? extends ElementMatcher<? super MethodDescription>, String> transformers() {
-    final Map<ElementMatcher<? super MethodDescription>, String> transformers = new HashMap<>();
-
-    transformers.put(
-        isMethod().and(returns(safeHasSuperType(named("org.hibernate.procedure.ProcedureCall")))),
-        GetProcedureCallAdvice.class.getName());
-
-    return transformers;
-  }
-
-  public static class GetProcedureCallAdvice {
-
-    @Advice.OnMethodExit(suppress = Throwable.class)
-    public static void getProcedureCall(
-        @Advice.This final SharedSessionContract session,
-        @Advice.Return final ProcedureCall returned) {
-
-      final ContextStore<SharedSessionContract, SessionState> sessionContextStore =
-          InstrumentationContext.get(SharedSessionContract.class, SessionState.class);
-      final ContextStore<ProcedureCall, SessionState> returnedContextStore =
-          InstrumentationContext.get(ProcedureCall.class, SessionState.class);
-
-      SessionMethodUtils.attachSpanFromStore(
-          sessionContextStore, session, returnedContextStore, returned);
-    }
-  }
-}
