--- dd-java-agent/instrumentation/hibernate-common/src/main/java/datadog/trace/instrumentation/hibernate/common/SessionState.java
@@ -1,13 +1,0 @@
-package datadog.trace.instrumentation.hibernate.common;
-
-import io.opentracing.Scope;
-import io.opentracing.Span;
-import lombok.Data;
-import lombok.NonNull;
-
-@Data
-public class SessionState {
-  @NonNull public Span sessionSpan;
-  public Scope methodScope;
-  public boolean hasChildSpan = true;
-}
