+++ dd-java-agent/instrumentation/hibernate/core-3.5/src/test/groovy/CriteriaTest.groovy
