+++ dd-java-agent/instrumentation/hibernate/core-3.5/src/test/resources/hibernate.cfg.xml
