+++ dd-java-agent/instrumentation/hibernate/core-4.0/src/main/java/datadog/trace/instrumentation/hibernate/core/v4_0/SessionFactoryInstrumentation.java
@@ -1,0 +1,82 @@
+package datadog.trace.instrumentation.hibernate.core.v4_0;
+
+import static datadog.trace.agent.tooling.ByteBuddyElementMatchers.safeHasSuperType;
+import static datadog.trace.instrumentation.hibernate.HibernateDecorator.DECORATOR;
+import static java.util.Collections.singletonMap;
+import static net.bytebuddy.matcher.ElementMatchers.isInterface;
+import static net.bytebuddy.matcher.ElementMatchers.isMethod;
+import static net.bytebuddy.matcher.ElementMatchers.named;
+import static net.bytebuddy.matcher.ElementMatchers.not;
+import static net.bytebuddy.matcher.ElementMatchers.returns;
+import static net.bytebuddy.matcher.ElementMatchers.takesArguments;
+
+import com.google.auto.service.AutoService;
+import datadog.trace.agent.tooling.Instrumenter;
+import datadog.trace.bootstrap.ContextStore;
+import datadog.trace.bootstrap.InstrumentationContext;
+import datadog.trace.instrumentation.hibernate.SessionState;
+import io.opentracing.Span;
+import io.opentracing.util.GlobalTracer;
+import java.util.Map;
+import net.bytebuddy.asm.Advice;
+import net.bytebuddy.description.method.MethodDescription;
+import net.bytebuddy.description.type.TypeDescription;
+import net.bytebuddy.matcher.ElementMatcher;
+import org.hibernate.SharedSessionContract;
+
+@AutoService(Instrumenter.class)
+public class SessionFactoryInstrumentation extends Instrumenter.Default {
+
+  public SessionFactoryInstrumentation() {
+    super("hibernate", "hibernate-core");
+  }
+
+  @Override
+  public Map<String, String> contextStore() {
+    return singletonMap("org.hibernate.SharedSessionContract", SessionState.class.getName());
+  }
+
+  @Override
+  public String[] helperClassNames() {
+    return new String[] {
+      "datadog.trace.instrumentation.hibernate.SessionState",
+      "datadog.trace.agent.decorator.BaseDecorator",
+      "datadog.trace.agent.decorator.ClientDecorator",
+      "datadog.trace.agent.decorator.DatabaseClientDecorator",
+      "datadog.trace.agent.decorator.OrmClientDecorator",
+      "datadog.trace.instrumentation.hibernate.HibernateDecorator",
+    };
+  }
+
+  @Override
+  public ElementMatcher<TypeDescription> typeMatcher() {
+    return not(isInterface()).and(safeHasSuperType(named("org.hibernate.SessionFactory")));
+  }
+
+  @Override
+  public Map<? extends ElementMatcher<? super MethodDescription>, String> transformers() {
+    return singletonMap(
+        isMethod()
+            .and(named("openSession").or(named("openStatelessSession")))
+            .and(takesArguments(0))
+            .and(
+                returns(
+                    named("org.hibernate.Session").or(named("org.hibernate.StatelessSession")))),
+        SessionFactoryAdvice.class.getName());
+  }
+
+  public static class SessionFactoryAdvice {
+
+    @Advice.OnMethodExit(suppress = Throwable.class)
+    public static void openSession(@Advice.Return final SharedSessionContract session) {
+
+      final Span span = GlobalTracer.get().buildSpan("hibernate.session").start();
+      DECORATOR.afterStart(span);
+      DECORATOR.onConnection(span, session);
+
+      final ContextStore<SharedSessionContract, SessionState> contextStore =
+          InstrumentationContext.get(SharedSessionContract.class, SessionState.class);
+      contextStore.putIfAbsent(session, new SessionState(span));
+    }
+  }
+}
