+++ dd-java-agent/instrumentation/hibernate/core-4.0/src/test/groovy/AbstractHibernateTest.groovy
