+++ dd-java-agent/instrumentation/hibernate/core-4.0/src/test/java/Value.java
