+++ dd-java-agent/instrumentation/hibernate/core-4.3/src/test/java/Value.java
