+++ dd-java-agent/instrumentation/hibernate/core-4.3/src/test/resources/hibernate.cfg.xml
