+++ dd-java-agent/instrumentation/hibernate/hibernate.gradle
