+++ dd-java-agent/instrumentation/hibernate/src/main/java/datadog/trace/instrumentation/hibernate/SessionState.java
@@ -1,0 +1,13 @@
+package datadog.trace.instrumentation.hibernate;
+
+import io.opentracing.Scope;
+import io.opentracing.Span;
+import lombok.Data;
+import lombok.NonNull;
+
+@Data
+public class SessionState {
+  @NonNull public Span sessionSpan;
+  public Scope methodScope;
+  public boolean hasChildSpan = true;
+}
