+++ settings.gradle
@@ -40,10 +40,10 @@ include ':dd-java-agent:instrumentation:elasticsearch:transport-5'
 include ':dd-java-agent:instrumentation:elasticsearch:transport-5.3'
 include ':dd-java-agent:instrumentation:elasticsearch:transport-6'
 include ':dd-java-agent:instrumentation:grpc-1.5'
-include ':dd-java-agent:instrumentation:hibernate-3.5'
-include ':dd-java-agent:instrumentation:hibernate-4.0'
-include ':dd-java-agent:instrumentation:hibernate-4.3'
-include ':dd-java-agent:instrumentation:hibernate-common'
+include ':dd-java-agent:instrumentation:hibernate'
+include ':dd-java-agent:instrumentation:hibernate:core-3.5'
+include ':dd-java-agent:instrumentation:hibernate:core-4.0'
+include ':dd-java-agent:instrumentation:hibernate:core-4.3'
 include ':dd-java-agent:instrumentation:http-url-connection'
 include ':dd-java-agent:instrumentation:hystrix-1.4'
 include ':dd-java-agent:instrumentation:jax-rs-annotations'
